isotptun: add -D option to daemonize after tun device created
System startup needs to wait for tun device to be created before
continuing on use of the tun device. Add -D option to fork a daemon of
isotptun and exit(EXIT_SUCCESS) when the tun device can be used.

Log errors to syslog when daemonized or to stderr when in foreground.

Signed-off-by: Marko Kohtala <marko.kohtala@gmail.com>