Updated README.md inspired by Alan (https://github.com/alan707)
Alan sent a pull request to update the README.md with some markdown tweaks
and a SocketCAN logo. Good idea!
After creating a SocketCAN logo PNG in the can-logos repository we can
reference the PNG inside the linux-can organization.
A license file needs to be added too - but this has to be discussed on the
mailing list first.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>