cansend: re-order socket handling and parsing of input
Do all the socket handling (setting address, bind, socket options) in
the beginning. Only parse the canframe argument after this is done.

This has as effect that wrong arguments are only reported after the
device has been opened and bound.

This is preparation for sending multiple CAN frames in one cansend
invocation.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>