can-utils: import PTX cansequence utility
The PTX flavour canutils[1] have a useful utility for generating a
sequence of CAN frames with an incrementing payload and checking such a
sequence for missed and reordered frames.

Recently, it has helped finding a regression in the kernel pfifo_fast
qdisc, which led to reordered frames[2].

To bring the utility to a greater audience, import the current state[3]
of it into the linux-can set of utilities.

[1]: https://git.pengutronix.de/cgit/tools/canutils
[2]: https://lore.kernel.org/netdev/661cc33a-5f65-2769-cc1a-65791cb4b131@pengutronix.de
[3]: 207a476b5 ("cansequence: print error frames in hex")

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>
[afa: code by mkl, I removed --version and fixed an off by one in strncpy]
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>