cangw: Update help text to indicate len8_dlc support
The can_dlc value that is passed to the kernel is not sanitized in cangw
therefore we only update the help text.

In the case that the CAN interface supports len8_dlc by setting
CAN_CTRLMODE_CC_LEN8_DLC the can-gw module can modify the full DLC value
range (0 .. 15) for Classic CAN frames.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>