circular buffer implementation
feature tries to co-exist with other options as best it can by
outputing a full log entry once before committing to a line size,
all the while padding the payload size to 25 characters.

So long as all other options are fixed width, this will work fine
for any future options as well.