canfdtest: Follow Bash exit status when signaled
Bash and many other shells use 128 + signum as the exit status when a
program get signal and exit.  Follow the common behavior so that we
know how the programs are killed.

canfdtest.c was effectively raising the received signal again to run
the default signal handler.  This has been declined by the maintainer
because it's over-engineering.  This commit replaces it with the
method proposed by the maintainer.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>