j1939acd: Follow Bash exit status when signaled
Bash and many other shells use 128 + signum as the exit status when a
program get signal and exit.  Follow the common behavior so that we
know how the programs are killed.

For j1939acd.c, it was using s.sig_term as a boolean key to exit from
the while loop.  This commit replaces a.sig_term with a.signal_num,
which holds the signal it received.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>