slcand: Follow Bash exit status when signaled
Bash and many other shells use 128 + signum as the exit status when a
program get signal and exit.  Follow the common behavior so that we
know how the programs are killed.

slcand was returning 0 when SIGINT and SIGTERM is received.  This
commit, instead, return 128 + signum when it gets a signal.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>