cansequence: main(): use C99 initializers for struct sockaddr_can addr
By using C99 initializers for struct sockaddr_can addr the not
assigned members are automatically set to 0x0, resulting in not
passing uninitialized memory to the bind() syscall.

Also remove the printing of the socket family, type and proto, this
was interesting back in the days, when the numbers were not fixed, yet.