cangen: allow to use -m on classical CAN interfaces
The -m option currently only works on CAN FD or CAN XL interfaces.
There is a logic to add CAN XL to the mix only if supported but CAN FD
is always forced.

Modify the -m logic so that only the options supported by the
interface are added to the mix. This way:

  - a Classical CAN interface only mixes -e and -R
  - a CAN FD interface mixes -e, -R, -f, -b and -E
  - a CAN XL interface mixes -e, -R, -f, -b, -E and -X

This provides a better user experience and also makes -m a good
default option for fuzzing any type of CAN interface.

Signed-off-by: Vincent Mailhol <mailhol@kernel.org>
Link: https://lore.kernel.org/r/20250907060330.441165-1-mailhol@kernel.org
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>